backend/requirements.txt
Django
django-cors-headers
djangorestframework
pylint-django
